# cores/api-nestjs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files": {
      "post": {
        "description": "Compatible `fetch + FormData` (web et React Native `{ uri, name, type }`). Le client NE force PAS `Content-Type: multipart/form-data` (boundary auto). Taille max = FILE_MAX_SIZE_BYTES ; types réels inspectés par signatures. Le client ne fournit ni ownerId ni clé de stockage.",
        "operationId": "files_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "category"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Contenu binaire (requis)."
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "IMAGE",
                      "DOCUMENT",
                      "AVATAR",
                      "MEDIA",
                      "VIDEO",
                      "AUDIO",
                      "IDENTITY_DOCUMENT",
                      "ATTACHMENT",
                      "OTHER"
                    ],
                    "description": "Catégorie générique (requise)."
                  },
                  "subjectId": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Rattachement métier générique optionnel."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fichier stocké et validé.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicStoredFileDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-06-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Multipart invalide / contenu non détecté / extension incohérente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.upload absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Quota propriétaire dépassé (nombre/octets).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "Fichier trop volumineux.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Trop de requêtes d’upload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Stockage objet indisponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload multipart d’un fichier privé (permission files.upload, ownership = utilisateur courant).",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}": {
      "get": {
        "operationId": "files_getMetadata",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Métadonnées publiques du fichier possédé (statut inclus).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PublicStoredFileDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-06-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identifiant invalide (UUID attendu).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.read absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Fichier absent ou non possédé (anti-énumération).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Métadonnées publiques d’un fichier possédé (permission files.read + ownership).",
        "tags": [
          "Files"
        ]
      },
      "delete": {
        "operationId": "files_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Fichier supprimé (objet S3 + marquage DELETED), idempotent."
          },
          "400": {
            "description": "Identifiant invalide (UUID attendu).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.delete absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Fichier absent ou non possédé (anti-énumération).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Finalisation DB impossible après suppression.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Stockage objet indisponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime un fichier (objet S3 puis DELETED ; permission files.delete + ownership ; idempotent).",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}/download-url": {
      "post": {
        "operationId": "files_createDownloadUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "URL signée temporaire (url + expiresAt). Réponse non mise en cache (no-store).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SignedDownloadResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-06-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identifiant invalide (UUID attendu).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.download absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Fichier absent ou non possédé (anti-énumération).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Fichier non téléchargeable (statut/visibilité).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Trop de demandes d’URL de téléchargement.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "Génération d’URL indisponible (objet manquant / signature).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Émet une URL signée de lecture courte (permission files.download + ownership ; non cacheable).",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}/quarantine": {
      "post": {
        "operationId": "files_quarantine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuarantineFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fichier mis en quarantaine (aucun accès/URL tant que la quarantaine dure).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-06-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identifiant invalide / raison hors valeurs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.quarantine absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Fichier introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Transition de statut invalide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met un fichier en quarantaine (administratif : permission files.quarantine, sans ownership).",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}/restore": {
      "post": {
        "operationId": "files_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quarantaine levée ; le propriétaire retrouve l’accès normal.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-06-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Identifiant invalide (UUID attendu).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentification requise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Permission files.restore absente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Fichier introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Transition de statut invalide (objet présent et empreinte d’intégrité requis).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lève la quarantaine (administratif : permission files.restore, sans ownership).",
        "tags": [
          "Files"
        ]
      }
    }
  },
  "info": {
    "title": "Enistere API Core NestJS",
    "description": "Contrat OpenAPI canonique du API Core NestJS V1 — source de vérité des API publiques (ADR-016).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PublicStoredFileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "b3f1c2d4-5e6f-4a8b-9c0d-1e2f3a4b5c6d"
          },
          "originalName": {
            "type": "string",
            "example": "photo.jpg",
            "description": "Métadonnée non fiable (UX uniquement)."
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "extension": {
            "type": "string",
            "example": "jpg"
          },
          "size": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "example": "1048576",
            "description": "Taille en octets (chaîne décimale, BigInt)."
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PRIVATE",
              "PUBLIC",
              "SIGNED",
              "INTERNAL"
            ],
            "example": "PRIVATE"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "UPLOADED",
              "VALIDATED",
              "REJECTED",
              "QUARANTINED",
              "DELETED"
            ],
            "example": "VALIDATED"
          },
          "category": {
            "type": "string",
            "enum": [
              "IMAGE",
              "DOCUMENT",
              "AVATAR",
              "MEDIA",
              "VIDEO",
              "AUDIO",
              "IDENTITY_DOCUMENT",
              "ATTACHMENT",
              "OTHER"
            ],
            "example": "IMAGE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-09T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-09T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "originalName",
          "mimeType",
          "extension",
          "size",
          "visibility",
          "status",
          "category",
          "createdAt",
          "updatedAt"
        ]
      },
      "SignedDownloadResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://storage.example.test/files/download?token=EXAMPLE-SIGNED-URL",
            "description": "URL temporaire signée (courte). Exemple fictif ; ne pas journaliser ni persister."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-09T12:05:00.000Z",
            "description": "Expiration de l’URL."
          }
        },
        "required": [
          "url",
          "expiresAt"
        ]
      },
      "QuarantineFileDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}